currency_converter/requirements.txt
# requirements.txt
Django>=3.2,<4.0
requests>=2.25,<3.0
python-decouple>=3.5,<4.0
